- fixed bug with colors in settings - add alerts to show errors - add class to handle internet connection
