--- Weather/Weather/Classes/Common/Services/Permissions/PermissionService.swift
@@ -1,13 +1,0 @@
-//
-//  PermissionService.swift
-//  Weather
-//
-//  Created by Alena Nesterkina on 24.12.20.
-//
-
-import Foundation
-
-protocol PermissionService: AnyObject {
-    func resolveStatus(for permission: Permission)
-    func request(permission: Permission)
-}
