+++ Weather/Weather/Classes/Common/Services/Reachability/Reachability.swift
@@ -1,0 +1,8 @@
+//
+//  Reachability.swift
+//  Weather
+//
+//  Created by Alena Nesterkina on 17.02.21.
+//
+
+import Foundation
