+++ Weather/Weather/Classes/Screens/MainScreen/Components/Chart/Components/IndicatorPoint.swift
