+++ Weather/Weather/Classes/Screens/MainScreen/Components/Chart/Components/MagnifierRect.swift
