+++ Weather/Weather/Classes/Screens/MainScreen/Components/DetailedView.swift
