+++ Weather/Weather/Classes/Screens/MainScreen/MainViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  MainViewModel.swift
+//  Weather
+//
+//  Created by Alena Nesterkina on 17.02.21.
+//
+
+import Foundation
